requirements.txt
python-dotenv
requests

# --- LLM / Gemini (OpenAI-compatible) ---
openai>=1.3.0

# --- MCP (multi-agent framework) ---
fastmcp
mcp

# --- Embedding 模型（BGE-M3）---
sentence-transformers
torch
transformers

# --- RAG 用到的工具 ---
numpy

# --- Agent Server 需要 ---
aiofiles
anyio
langchain_community
faiss-cpu
pandas

# --- 資料庫連線 (PostgreSQL 目前沒有用到) ---
# psycopg2-binary>=2.9.0
# asyncpg>=0.29.0

# --- 其他工具 ---
# 如果未來要加上網路搜尋或其他功能，可以在此擴充
